search layouts

+++ JewellBiz/src/com/jewellbiz/android/jewellbiz/data/JbzDatabase.java
@@ -5,12 +5,13 @@ import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 import android.database.sqlite.SQLiteQueryBuilder;
+import android.provider.BaseColumns;
 import android.util.Log;
 
 public class JbzDatabase extends SQLiteOpenHelper {
 	
 	private static final String DEBUG_TAG = "JbzDatabase";
-	public static final int DB_VERSION = 1;
+	public static final int DB_VERSION = 4;
 	private static final String DB_NAME = "jbz_database";
 	
 	public static final String TABLE_ARTICLES = "articles";
@@ -19,21 +20,24 @@ public class JbzDatabase extends SQLiteOpenHelper {
 	public static final String COL_URL = "url";
 	
 	public static final String COL_DATE = "article_date";
-	private static final String ALTER_ADD_COL_DATE = "ALTER TABLE "
-			+ TABLE_ARTICLES + " ADD COLUMN " + COL_DATE 
-			+ " INTEGER NOT NULL DEFAULT '1297728000' ";
 	
 	public static final String COL_READ = "read";
-	private static final String ALTER_ADD_COL_READ = "ALTER TABLE "
-			+ TABLE_ARTICLES + " ADD COLUMN " + COL_READ
-			+ " INTEGER NOT NULL DEFAULT 0";
+	
+	public static final String COL_DESC = "desc";
 	
 	private static final String CREATE_TABLE_ARTICLES = "CREATE TABLE " 
 			+ TABLE_ARTICLES + " (" + ID 
 			+ " integer PRIMARY KEY AUTOINCREMENT, " + COL_TITLE 
 			+ " text NOT NULL, " + COL_URL + " text UNIQUE NOT NULL, "
-			+ COL_DATE + " INTEGER NOT NULL DEFAULT (strftime('%s','now')), "	
+			+ COL_DATE + " INTEGER NOT NULL DEFAULT (strftime('%s','now')), "
+			+ COL_DESC + " text, "
 			+ COL_READ + " INTEGER NOT NULL default 0" + ");";
+	
+	public static final String TABLE_ARTICLES_SEARCH = "searchtable";
+	public static final String _DOCID = "docid";
+	private static final String CREATE_TABLE_ARTICLE_SEARCH = "CREATE VIRTUAL TABLE "
+			+ TABLE_ARTICLES_SEARCH + " USING FTS3(" + COL_TITLE + " TEXT, " 
+			+ COL_DESC + " TEXT" + ");";
 	
 	private static final String DB_SCHEMA = CREATE_TABLE_ARTICLES;
 	
@@ -44,34 +48,76 @@ public class JbzDatabase extends SQLiteOpenHelper {
 	@Override
 	public void onCreate(SQLiteDatabase db) {
 		db.execSQL(DB_SCHEMA);
+		db.execSQL(CREATE_TABLE_ARTICLE_SEARCH);
 		seedData(db);
+		
 	}
 	
 
 	@Override
 	public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-		Log.w(DEBUG_TAG, "Upgrading database");
-		db.execSQL("DROP TABLE IF EXISTS " + TABLE_ARTICLES);
-		onCreate(db);
+		Log.d(DEBUG_TAG, "onUpgrade() from " + oldVersion + " to " + newVersion);
 		
-		// example of how to check for when upgrading database
-		// for new app versions
-		/*if (newVersion == 3) {
-			// try to keep the date, using alter tables
-			if (oldVersion == 2) {
-				db.execSQL(ALTER_ADD_COL_READ);
-			} else if (oldVersion == 1) {
-				db.execSQL(ALTER_ADD_COL_DATE);
-				db.execSQL(ALTER_ADD_COL_READ);
-			}
-		} else {
-			Log.w(DEBUG_TAG, "Upgrading database. Existing contents will be lost. ["
-	                + oldVersion + "]->[" + newVersion + "]");
-	        db.execSQL("DROP TABLE IF EXISTS " + TABLE_POSTS);
-	        onCreate(db);
-		} */
+		// run necessary upgrades
+        int version = oldVersion;
+        switch (version) {
+        case 1:
+        	upgradeToTwo(db);
+        	version = 2;
+        case 2:
+        	upgradeToThree(db);
+        	version = 3;
+        case 3:
+        	upgradeToFour(db);
+        	version = 4;
+        }
+        
+        // drop all tables if version is not right
+        Log.d(DEBUG_TAG, "after upgrade logic, at version " + version);
+        if (version != DB_VERSION) {
+        	Log.w(DEBUG_TAG, "Database has incompatible version, starting from scratch");
+        	db.execSQL("DROP TABLE IF EXISTS " + TABLE_ARTICLES);
+        	db.execSQL("DROP TABLE IF EXISTS " + TABLE_ARTICLES_SEARCH);
+    		onCreate(db);
+        }
 		
 	}
+	
+	/**
+	 * Add a {@link Read} column for marking articles as read
+	 */
+	private void upgradeToFour(SQLiteDatabase db) {
+		db.execSQL("ALTER TABLE " + TABLE_ARTICLES + " ADD COLUMN " + COL_DESC + " text");
+	}
+	
+	/**
+	 * Add a {@link Read} column for marking articles as read
+	 */
+	private void upgradeToThree(SQLiteDatabase db) {
+		db.execSQL("ALTER TABLE " + TABLE_ARTICLES + " ADD COLUMN " + COL_READ + " INTEGER NOT NULL DEFAULT 0");
+	}
+	
+	/**
+	 * Add a {@link Date} column for storing article date
+	 */
+	private void upgradeToTwo(SQLiteDatabase db) {
+		db.execSQL("ALTER TABLE " + TABLE_ARTICLES + " ADD COLUMN " + COL_DATE + " INTEGER NOT NULL DEFAULT '1297728000' ");
+	}
+	
+	public static void onRenewFTSTable(SQLiteDatabase db) {
+        db.beginTransaction();
+        try {
+            db.execSQL("drop table if exists " + TABLE_ARTICLES_SEARCH);
+            db.execSQL(CREATE_TABLE_ARTICLE_SEARCH);
+            db.execSQL("INSERT INTO " + TABLE_ARTICLES_SEARCH + "(docid," + COL_TITLE + ","
+                    + COL_DESC + ")" + " select " + ID + "," + COL_TITLE
+                    + "," + COL_DESC + " from " + TABLE_ARTICLES + ";");
+            db.setTransactionSuccessful();
+        } finally {
+            db.endTransaction();
+        }
+    }
+	
 	
 	private void seedData(SQLiteDatabase db) {
         db.execSQL("insert into articles (title, url, article_date) values ('Welcome to the Spring Edition of Jewell Business Today', 'http://jewellbiz.com/2012/03/15/welcome-to-the-spring-edition-of-jewell-business-today/', (strftime('%s', '2012-03-15')));");
@@ -82,27 +128,71 @@ public class JbzDatabase extends SQLiteOpenHelper {
     }
 	
     
-    public Cursor getWordMatches(String query, String[] columns) {
-    	String selection = COL_TITLE + " MATCH ?";
-    	String[] selectionArgs = new String[] {query+"*"};
-    	
-		return query(selection, selectionArgs, columns);
-    }
+	public static Cursor search(String selection, String[] selectionArgs, SQLiteDatabase db) {
+		
+		StringBuilder query = new StringBuilder();
+		// select final result columns
+		query.append("SELECT ");
+        query.append(ID).append(",");
+        query.append(COL_TITLE).append(",");
+        query.append(COL_DESC).append(",");
+        
+        query.append(" FROM ");
+        query.append("(");
 
-	private Cursor query(String selection, String[] selectionArgs, String[] columns) {
-		SQLiteQueryBuilder builder = new SQLiteQueryBuilder();
-		builder.setTables(TABLE_ARTICLES);
+        // join all shows...
+        query.append("(");
+        query.append("SELECT ").append(BaseColumns._ID).append(" as sid,").append(COL_TITLE);
+        query.append(" FROM ").append(TABLE_ARTICLES);
+        query.append(")");
+        
+        query.append(" JOIN ");
+
+        // ...with matching episodes
+        query.append("(");
+        query.append("SELECT ");
+        query.append(ID).append(",");
+        query.append(COL_TITLE).append(",");
+        query.append(COL_DESC).append(",");
+
+        query.append(" FROM ");
+        // join searchtable results...
+        query.append("(");
+        query.append("SELECT ");
+        query.append(_DOCID).append(",");
+        query.append("snippet(" + TABLE_ARTICLES_SEARCH + ",'<b>','</b>','...')").append(" AS ")
+                .append(COL_DESC);
+        query.append(" FROM ").append(TABLE_ARTICLES_SEARCH);
+        query.append(" WHERE ").append(TABLE_ARTICLES_SEARCH).append(" MATCH ?");
+        query.append(")");
+        query.append(" JOIN ");
+        // ...with episodes table
+        query.append("(");
+        query.append("SELECT ");
+        query.append(ID).append(",");
+        query.append(COL_TITLE).append(",");
+        query.append(COL_DESC).append(",");
+        query.append(" FROM ").append(TABLE_ARTICLES);
+        query.append(")");
+        query.append(" ON ").append(ID).append("=").append(_DOCID);
+
+        query.append(")");
+        
+     // append given selection
+        if (selection != null) {
+            query.append(" WHERE ");
+            query.append("(").append(selection).append(")");
+        }
+        
+     // ordering
+        query.append(" ORDER BY ");
+        query.append(COL_TITLE).append(" ASC,");
+        
+     // search for anything starting with the given search term
+        selectionArgs[0] = "\"" + selectionArgs[0] + "*\"";
+
+        return db.rawQuery(query.toString(), selectionArgs);
 		
-		Cursor cursor = builder.query(getReadableDatabase(), columns, selection, selectionArgs, null, null, null);
-		
-		if (cursor == null) {
-			return null;
-		} else if (!cursor.moveToFirst()) {
-			cursor.close();
-			return null;
-		}
-		
-		return cursor;
 	}
 
 }
